Editor: Unify the sidebar between the post and site editors
