+++ packages/edit-site/src/components/global-styles-sidebar/style.scss
