--- packages/edit-site/src/components/sidebar-edit-mode/settings-header/style.scss
@@ -1,13 +1,0 @@
-.components-panel__header.edit-site-sidebar-edit-mode__panel-tabs {
-	padding-left: 0;
-
-	.components-button.has-icon {
-		padding: 0;
-		min-width: $icon-size;
-		height: $icon-size;
-
-		@include break-medium() {
-			display: flex;
-		}
-	}
-}
