--- packages/edit-site/src/components/sidebar-edit-mode/template-panel/style.scss
@@ -1,16 +1,0 @@
-.edit-site-template-card {
-	&__actions {
-		line-height: 0;
-		> .components-button.is-small.has-icon {
-			padding: 0;
-			min-width: auto;
-		}
-		flex-shrink: 0;
-	}
-}
-
-
-.edit-site-template-panel .block-editor-block-preview__container {
-	border-radius: 2px;
-	box-shadow: none;
-}
