+++ packages/editor/src/components/sidebar/index.js
@@ -13,11 +13,9 @@ import {
 	useEffect,
 	useRef,
 } from '@wordpress/element';
-import { isRTL, __, sprintf } from '@wordpress/i18n';
+import { isRTL, __ } from '@wordpress/i18n';
 import { drawerLeft, drawerRight } from '@wordpress/icons';
 import { store as keyboardShortcutsStore } from '@wordpress/keyboard-shortcuts';
-import { addQueryArgs } from '@wordpress/url';
-import { store as noticesStore } from '@wordpress/notices';
 import { privateApis as componentsPrivateApis } from '@wordpress/components';
 import { store as interfaceStore } from '@wordpress/interface';
 
@@ -34,7 +32,9 @@ import PostDiscussionPanel from '../post-discussion/panel';
 import PostLastRevisionPanel from '../post-last-revision/panel';
 import PostSummary from './post-summary';
 import PostTaxonomiesPanel from '../post-taxonomies/panel';
+import PostTransformPanel from '../post-transform-panel';
 import SidebarHeader from './header';
+import TemplateContentPanel from '../template-content-panel';
 import useAutoSwitchEditorSidebars from '../provider/use-auto-switch-editor-sidebars';
 import { sidebars } from './constants';
 import { unlock } from '../../lock-unlock';
@@ -51,6 +51,8 @@ const SidebarContent = ( {
 	tabName,
 	keyboardShortcut,
 	isEditingTemplate,
+	renderingMode,
+	onActionPerformed,
 	extraPanels,
 } ) => {
 	const tabListRef = useRef( null );
@@ -85,56 +87,6 @@ const SidebarContent = ( {
 			selectedTabElement?.focus();
 		}
 	}, [ tabName ] );
-	const { createSuccessNotice } = useDispatch( noticesStore );
-
-	const onActionPerformed = useCallback(
-		( actionId, items ) => {
-			switch ( actionId ) {
-				case 'move-to-trash':
-					{
-						const postType = items[ 0 ].type;
-						document.location.href = addQueryArgs( 'edit.php', {
-							post_type: postType,
-						} );
-					}
-					break;
-				case 'duplicate-post':
-					{
-						const newItem = items[ 0 ];
-						const title =
-							typeof newItem.title === 'string'
-								? newItem.title
-								: newItem.title?.rendered;
-						createSuccessNotice(
-							sprintf(
-								// translators: %s: Title of the created post e.g: "Post 1".
-								__( '"%s" successfully created.' ),
-								title
-							),
-							{
-								type: 'snackbar',
-								id: 'duplicate-post-action',
-								actions: [
-									{
-										label: __( 'Edit' ),
-										onClick: () => {
-											const postId = newItem.id;
-											document.location.href =
-												addQueryArgs( 'post.php', {
-													post: postId,
-													action: 'edit',
-												} );
-										},
-									},
-								],
-							}
-						);
-					}
-					break;
-			}
-		},
-		[ createSuccessNotice ]
-	);
 
 	return (
 		<PluginSidebar
@@ -167,6 +119,10 @@ const SidebarContent = ( {
 					/>
 					{ ! isEditingTemplate && <PostSummary /> }
 					<PluginDocumentSettingPanel.Slot />
+					{ renderingMode !== 'post-only' && (
+						<TemplateContentPanel />
+					) }
+					<PostTransformPanel />
 					<PostLastRevisionPanel />
 					<PostTaxonomiesPanel />
 					<PostDiscussionPanel />
@@ -182,10 +138,10 @@ const SidebarContent = ( {
 	);
 };
 
-const Sidebar = ( { extraPanels } ) => {
+const Sidebar = ( { extraPanels, onActionPerformed } ) => {
 	useAutoSwitchEditorSidebars();
-	const { tabName, keyboardShortcut, isEditingTemplate } = useSelect(
-		( select ) => {
+	const { tabName, keyboardShortcut, isEditingTemplate, renderingMode } =
+		useSelect( ( select ) => {
 			const shortcut = select(
 				keyboardShortcutsStore
 			).getShortcutRepresentation( 'core/editor/toggle-sidebar' );
@@ -211,10 +167,9 @@ const Sidebar = ( { extraPanels } ) => {
 				isEditingTemplate:
 					select( editorStore ).getCurrentPostType() ===
 					'wp_template',
+				renderingMode: select( editorStore ).getRenderingMode(),
 			};
-		},
-		[]
-	);
+		}, [] );
 
 	const { enableComplementaryArea } = useDispatch( interfaceStore );
 
@@ -237,6 +192,8 @@ const Sidebar = ( { extraPanels } ) => {
 				tabName={ tabName }
 				keyboardShortcut={ keyboardShortcut }
 				isEditingTemplate={ isEditingTemplate }
+				renderingMode={ renderingMode }
+				onActionPerformed={ onActionPerformed }
 				extraPanels={ extraPanels }
 			/>
 		</Tabs>
